Added All Player Data Sync
IMPORTANT: This needs to be called after character selection is completed. Not entirely sure how this works so not added yet.

Added in sync for relevant information about other players (their IDs, names and character types) to be accessed on the client. This means that only the player ID has to be sent when display other characters info.