Test Fix for Player Death
Not working as it keeps skipping over players in the turn order when incrementing active player now. Not sure why, but I think its something to do with IsDead property. I did change a lot of how things work since I think this would be a better way of handling things, however if you cant figure it out feel free to revert it